+++ tests/core/mod.rs
@@ -1,2 +1,2 @@
-pub mod action;
+pub mod config;
 pub mod postit;
