+++ tests/mod.rs
@@ -1,4 +1,3 @@
-pub mod config;
 pub mod core;
 pub mod docs;
 pub mod mocks;
